+++ src/App.tsx
@@ -1,14 +1,11 @@
 import { useEffect } from 'react';
 import { CssVarsProvider } from '@mui/joy/styles';
 import CssBaseline from '@mui/joy/CssBaseline';
-import Box from '@mui/joy/Box';
-import Sidebar from './common/Sidebar';
-import Header from './common/Header';
-import MyMessages from './chat/MyMessages';
 import { Route, Routes, Navigate } from 'react-router-dom';
 import { customTheme } from './style';
 import { useSocket } from './state/chat';
 import UserJoin from './pages/user/UserJoin';
+import { ChatMain } from './pages/chat/ChatMain';
 
 export default function App() {
   const connectSocket = useSocket((state) => state.connectSocket);
@@ -23,21 +20,9 @@ export default function App() {
       <Routes>
         <Route path="/" element={<Navigate replace to="/user" />} />
         <Route path="/user" element={<UserJoin />} />
-        <Route path="/chat" element={<Dashboard />} />
-        <Route path="/chat/:id" element={<Dashboard />} />
+        <Route path="/chat" element={<ChatMain />} />
+        <Route path="/chat/:id" element={<ChatMain />} />
       </Routes>
     </CssVarsProvider>
-  );
-}
-
-function Dashboard() {
-  return (
-    <Box sx={{ display: 'flex', minHeight: '100dvh' }}>
-      <Sidebar />
-      <Header />
-      <Box component="main" className="MainContent" sx={{ flex: 1 }}>
-        <MyMessages />
-      </Box>
-    </Box>
   );
 }
