+++ src/components/chat/AiPannel.tsx
