+++ src/components/chat/ChatBubble.tsx
@@ -8,9 +8,9 @@ import Typography from '@mui/joy/Typography';
 import CelebrationOutlinedIcon from '@mui/icons-material/CelebrationOutlined';
 import FavoriteBorderIcon from '@mui/icons-material/FavoriteBorder';
 import InsertDriveFileRoundedIcon from '@mui/icons-material/InsertDriveFileRounded';
-import { MessageProps } from '../types';
+import { TMessageProps } from '../../pages/chat/type';
 
-type ChatBubbleProps = MessageProps & {
+type ChatBubbleProps = TMessageProps & {
   variant: 'sent' | 'received';
 };
