+++ src/components/common/AvatarWithStatus.tsx
