+++ src/pages/user/type.ts
@@ -1,0 +1,6 @@
+export type TUserProps = {
+  name: string;
+  username: string;
+  avatar: string;
+  online: boolean;
+};
